+++ packages/terra-infinite-list/tests/wdio/infinite-list-spec.js
@@ -53,4 +53,19 @@ describe('InfiniteList', () => {
 
     Terra.should.matchScreenshot({ selector: '#test-infinite-list' });
   });
+
+  describe('Displays an infinite list same count and different content', () => {
+    before(() => {
+      browser.setViewportSize(Terra.viewports('small')[0]);
+      browser.url('/#/raw/tests/infinite-list/infinite-list-same-count');
+      browser.waitForVisible('#test-infinite-list');
+      browser.pause(150);
+    });
+    Terra.should.matchScreenshot('before-update', { selector: '#test-infinite-list' });
+    it('update the child items', () => {
+      browser.click('#test-click');
+      browser.pause(50);
+    });
+    Terra.should.matchScreenshot('after-update', { selector: '#test-infinite-list' });
+  });
 });