+++ packages/terra-infinite-list/CHANGELOG.md
@@ -6,6 +6,10 @@ Unreleased
 ### Changed
 * Moved examples to src
 
+### Fixed
+* New props of children with different content update correctly
+* Double request for items prevented during render
+
 1.2.0 - (May 22, 2018)
 ------------------
 ### Changed